+++ src/chanjo2/main.py
@@ -4,9 +4,9 @@ import os
 import uvicorn
 from chanjo2 import __version__
 from chanjo2.dbutil import engine
-from chanjo2.demo.populate_demo import load_demo_data
 from chanjo2.endpoints import cases, intervals, samples
 from chanjo2.models.sql_models import Base
+from chanjo2.populate_demo import load_demo_data
 from fastapi import FastAPI, status
 
 LOG = logging.getLogger("uvicorn.access")
